[rocprofiler-compute] Route split-out tests to the module they exercise
The earlier splits placed tests by source file rather than by the symbol
each one calls, so two files held tests for modules they are not named
after.

- Move the file_lock tests to test_utils_profile.py; only the
  gpu_benchmark_lock wrapper test still belongs in test_benchmark_base.py
- Move the MIGPUSpecs tests to test_mi_gpu_spec.py, canonical_config_arch
  to test_utils_common.py, and is_single_panel_config to test_file_io.py,
  leaving test_specs.py covering utils/specs.py only

Bodies are unchanged, so node IDs are identical before and after.
