+++ projects/rocprofiler-compute/tests/test_utils_profile.py
@@ -3,6 +3,7 @@
 
 """Unit tests for utils/utils_profile.py."""
 
+import os
 from unittest import mock
 
 import pandas as pd
@@ -810,3 +811,99 @@ def test_get_submodules_docstring_verification(mock_walk, mock_import):
     assert isinstance(result, list)
     assert "submodule1" in result
     assert "submodule2" in result
+
+
+@pytest.mark.misc
+def test_file_lock_creates_world_rw_file(tmp_path):
+    """A freshly created lock file must be world-rw (0o666) regardless of umask."""
+    import stat
+
+    from utils import utils_profile
+
+    lock_file = tmp_path / "shared.lock"
+
+    # Force a strict umask that would otherwise leave the file owner-only.
+    old_umask = os.umask(0o077)
+    try:
+        with utils_profile.file_lock(lock_file):
+            assert lock_file.exists()
+    finally:
+        os.umask(old_umask)
+
+    file_mode = stat.S_IMODE(os.stat(lock_file).st_mode)
+    assert file_mode == 0o666, (
+        f"Lock file must be world-rw so any user can acquire it; got "
+        f"{oct(file_mode)}. A non-0o666 lock file locks out other users."
+    )
+
+
+@pytest.mark.misc
+def test_file_lock_does_not_change_process_umask(tmp_path, monkeypatch):
+    """Lock creation must not change process-global umask."""
+
+    def fail_if_called(_mask):
+        raise AssertionError("file_lock must not call os.umask()")
+
+    monkeypatch.setattr(utils_profile.os, "umask", fail_if_called)
+
+    with utils_profile.file_lock(tmp_path / "shared.lock"):
+        pass
+
+
+@pytest.mark.misc
+def test_file_lock_existing_file_owned_by_other_user(tmp_path, monkeypatch):
+    """A lock file owned by another user (no write access) is still lockable."""
+
+    from utils import utils_profile
+
+    lock_file = tmp_path / "shared.lock"
+    # Pre-create the lock file (as if another user created it first).
+    lock_file.touch()
+
+    real_os_open = os.open
+    opened_modes = []
+
+    def fake_os_open(path, flags, *args):
+        if flags & os.O_EXCL:
+            # Let the create-only attempt fail naturally (file exists).
+            return real_os_open(path, flags, *args)
+        if flags & os.O_RDWR:
+            opened_modes.append("rw")
+            raise PermissionError(13, "Permission denied")
+        opened_modes.append("ro")
+        return real_os_open(path, flags, *args)
+
+    monkeypatch.setattr(utils_profile.os, "open", fake_os_open)
+
+    acquired = False
+    with utils_profile.file_lock(lock_file):
+        acquired = True
+
+    assert acquired, "Lock must be acquired via read-only fallback"
+    assert opened_modes == ["rw", "ro"], (
+        "Should attempt read-write first, then fall back to read-only"
+    )
+
+
+@pytest.mark.misc
+def test_file_lock_unopenable_file_raises(tmp_path, monkeypatch):
+    """If the lock file cannot be opened at all, raise an actionable error."""
+
+    from utils import utils_profile
+
+    lock_file = tmp_path / "shared.lock"
+    lock_file.touch()
+
+    real_os_open = os.open
+
+    def fake_os_open(path, flags, *args):
+        if flags & os.O_EXCL:
+            # Let the create-only attempt fail naturally (file exists).
+            return real_os_open(path, flags, *args)
+        raise PermissionError(13, "Permission denied")
+
+    monkeypatch.setattr(utils_profile.os, "open", fake_os_open)
+
+    with pytest.raises(RuntimeError, match="Cannot open lock file"):
+        with utils_profile.file_lock(lock_file):
+            pass
